GEODE-3539: Refactor to put the alter runtime tests in a separate class
